requirement.txt
streamlit
easyocr
ultralytics
pillow